Abandon unit-cell model
Previous model assumes horizontal unit cell, and folded BZ. Such
interpretation is wrong because that cannot be realized on a lattice.
This lattice model actually contains the sublattice on the same site.
The unit cell is a single site. And the BZ should not be folded. This
commit correct this mistake.